+++ DatadogRUM/Sources/RUMMonitor/Scopes/RUMResourceScope.swift
@@ -276,11 +276,14 @@ internal class RUMResourceScope: RUMScope {
             device: .init(context: context, telemetry: dependencies.telemetry),
             display: nil,
             error: .init(
+                binaryImages: nil,
+                category: .exception, // resource errors are categorised as "Exception"
                 handling: nil,
                 handlingStack: nil,
                 id: nil,
                 isCrash: false,
                 message: command.errorMessage,
+                meta: nil,
                 resource: .init(
                     method: resourceHTTPMethod,
                     provider: errorEventProvider,
@@ -290,8 +293,11 @@ internal class RUMResourceScope: RUMScope {
                 source: command.errorSource.toRUMDataFormat,
                 sourceType: command.errorSourceType,
                 stack: command.stack,
-                type: command.errorType
+                threads: nil,
+                type: command.errorType,
+                wasTruncated: nil
             ),
+            freeze: nil,
             os: .init(context: context),
             service: context.service,
             session: .init(