+++ DatadogRUM/Tests/Instrumentation/RUMInstrumentationTests.swift
@@ -18,7 +18,11 @@ class RUMInstrumentationTests: XCTestCase {
             uiKitRUMViewsPredicate: UIKitRUMViewsPredicateMock(),
             uiKitRUMActionsPredicate: nil,
             longTaskThreshold: nil,
-            dateProvider: SystemDateProvider()
+            appHangThreshold: .mockAny(),
+            mainQueue: .main,
+            dateProvider: SystemDateProvider(),
+            backtraceReporter: BacktraceReporterMock(),
+            telemetry: NOPTelemetry()
         )
 
         // Then
@@ -37,7 +41,11 @@ class RUMInstrumentationTests: XCTestCase {
             uiKitRUMViewsPredicate: nil,
             uiKitRUMActionsPredicate: UIKitRUMActionsPredicateMock(),
             longTaskThreshold: nil,
-            dateProvider: SystemDateProvider()
+            appHangThreshold: .mockAny(),
+            mainQueue: .main,
+            dateProvider: SystemDateProvider(),
+            backtraceReporter: BacktraceReporterMock(),
+            telemetry: NOPTelemetry()
         )
 
         // Then
@@ -53,7 +61,11 @@ class RUMInstrumentationTests: XCTestCase {
             uiKitRUMViewsPredicate: nil,
             uiKitRUMActionsPredicate: nil,
             longTaskThreshold: 0.5,
-            dateProvider: SystemDateProvider()
+            appHangThreshold: .mockAny(),
+            mainQueue: .main,
+            dateProvider: SystemDateProvider(),
+            backtraceReporter: BacktraceReporterMock(),
+            telemetry: NOPTelemetry()
         )
 
         // Then
@@ -72,7 +84,11 @@ class RUMInstrumentationTests: XCTestCase {
             uiKitRUMViewsPredicate: nil,
             uiKitRUMActionsPredicate: nil,
             longTaskThreshold: .mockRandom(min: -100, max: 0),
-            dateProvider: SystemDateProvider()
+            appHangThreshold: .mockAny(),
+            mainQueue: .main,
+            dateProvider: SystemDateProvider(),
+            backtraceReporter: BacktraceReporterMock(),
+            telemetry: NOPTelemetry()
         )
 
         // Then
@@ -87,7 +103,11 @@ class RUMInstrumentationTests: XCTestCase {
             uiKitRUMViewsPredicate: UIKitRUMViewsPredicateMock(),
             uiKitRUMActionsPredicate: UIKitRUMActionsPredicateMock(),
             longTaskThreshold: 0.5,
-            dateProvider: SystemDateProvider()
+            appHangThreshold: .mockAny(),
+            mainQueue: .main,
+            dateProvider: SystemDateProvider(),
+            backtraceReporter: BacktraceReporterMock(),
+            telemetry: NOPTelemetry()
         )
         let subscriber = RUMCommandSubscriberMock()