+++ DatadogSessionReplay/Tests/Processor/ResourceProcessorTests.swift
@@ -62,4 +62,36 @@ class ResourceProcessorTests: XCTestCase {
 
         XCTAssertTrue(writer.resources.isEmpty)
     }
+
+    func testItRemovedDuplicates() {
+        let writer = ResourceWriterMock()
+        let processor = ResourceProcessor(
+            queue: NoQueue(),
+            resourcesWriter: writer
+        )
+
+        let resource1: MockResource = .mockRandom()
+        let resource2: MockResource = .mockRandom()
+        let context: EnrichedResource.Context = .mockRandom()
+
+        processor.process(resources: [resource1, resource2], context: context)
+        processor.process(resources: [resource1, resource2], context: context)
+
+        XCTAssertEqual(writer.resources.count, 1)
+        XCTAssertEqual(
+            writer.resources[0],
+            [
+                EnrichedResource(
+                    identifier: resource1.calculateIdentifier(),
+                    data: resource1.calculateData(),
+                    context: context
+                ),
+                EnrichedResource(
+                    identifier: resource2.calculateIdentifier(),
+                    data: resource2.calculateData(),
+                    context: context
+                ),
+            ]
+        )
+    }
 }